# doc/swagger/swagger.yaml
definitions:
  controller.Response:
    properties:
      data: {}
      error:
        example: ""
        type: string
    type: object
...
  input.ProductDto:
    properties:
      category_id:
        type: integer
      description:
        type: string
      price:
        type: number
      product_name:
        type: string
    required:
    - category_id
    - description
    - price
    - product_name
    type: object
...
  output.ProductDto:
    properties:
      active:
        type: boolean
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      price:
        type: number
      product_name:
        type: string
    type: object
  swagger.InternalServerErrorResponse:
    properties:
      data: {}
      error:
        example: Internal server error
        type: string
    type: object
...
info:
  contact: {}
paths:
  /categories/{id}/products:
    get:
      consumes:
      - application/json
      description: List products from a certain category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/output.ProductDto'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/swagger.InternalServerErrorResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: List product by category
      tags:
      - Products
...
  /products:
    post:
      consumes:
      - application/json
      description: Place a new product
      parameters:
      - description: Product properties
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/input.ProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                data:
                  $ref: '#/definitions/output.ProductDto'
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/swagger.InternalServerErrorResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Create product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Remove product from list
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/swagger.InternalServerErrorResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Delete product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update product properties
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product properties
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/input.ProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                data:
                  $ref: '#/definitions/output.ProductDto'
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/swagger.InternalServerErrorResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update product
      tags:
      - Products
swagger: "2.0"